BURGUNDYSIP WINE STORE ANALYSIS & MODELS
This project was performed as part of the Data Handling and Decision making Course in the Data Analytics Program at the Toronto School of Management as a final group project. 

Synopsis:
BurgundySip is a successful wine cellar/store that promotes high-quality wines from all over the world. BurgundySip management recently decided to use data analytics to streamline their pricing strategy. As a team of analysts, conduct a pilot exploratory (but still prescriptive) study to investigate data relationships, trends, and patterns in price as well as other variables.